GeneratorPQ classes no longer exist
Similar to commit 00d9e7c13493eaad7d4e2c03c39b548aa65ec5a8
Via the DSL part, we want to instanciate potential multiple GeneratorConnectedToOmegaRef, with different libs (first one is GeneratorPQ)
