+++ dynawaltz-dsl/src/main/resources/synchronous_generators.cfg
@@ -4,4 +4,8 @@ synchronousGeneratorsLibs {
      GeneratorSynchronousFourWindingsProportionalRegulationsStepPm
      GeneratorSynchronousThreeWindings
      GeneratorSynchronousThreeWindingsProportionalRegulations
- }
+}
+
+connectedToOmegaRefGeneratorsLibs {
+    GeneratorPQ
+}
